requirements.txt
Flask_PyMongo==2.3.0
Flask==1.1.2
pymongo==3.11.0
Pillow==8.0.1
protobuf==3.14.0
python-dotenv==0.15.0
